envision_api/requirements.txt
anyio==3.6.2
attrs==22.2.0
bcrypt==4.0.1
boto3==1.26.31
botocore==1.29.31
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
coverage==7.1.0
dnspython==2.2.1
exceptiongroup==1.1.0
fastapi==0.88.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
mailchimp-transactional==1.0.50
openai==0.27.4
packaging==23.0
passlib==1.7.4
Pillow==9.3.0
pluggy==1.0.0
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
pymongo==4.3.3
pytest==7.2.1
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.7
requests==2.28.1
rfc3986==1.5.0
s3transfer==0.6.0
sentinels==1.0.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
